I'm going to upload this for today
A little side thing I had started today, it's not going to take much more effort, except a lot of typing and then testing to make sure it doesn't mess anything up on my PC or anyone else. I'll leave it at that, so if someone is looking at it they can just make assumptions on all the random commented out stuff. It runs as it is now at least. About to work on that game, and the spreites should work, I had a problem with Python 3.7 because of the Marketplace installer, but fixxed that and sprites should work after adding frame rate really quick to yesterdays post.