Add all images to same directory as the gameclean.py File
Early upload, probably an extra, however I have sprite action in this pygame script. I had to use a Sailor moon sprite because people don't like posting transparency online. I didn't want to spend time making them work, adding transparent backgrounds, especially because I don't know what I am going to be using in the final stages. I will probably make my own sprites, maybe look at some sprite and pixel patterns to help get an idea on how certain styles are done, I want it to look good, also the walk up and down is not activated in this code, but it exists, and so do the images. It will be up soon, I also am going to change the jump into something else, maybe a dodge, and have two of them, for left and right. Jumping doesn't do as much in an RPG/topdown game, witch this will be, I feel like to many people make side scrollers, and it's just to easy to make a game have obstacles with platforming, I want to have to build complex things to interact with on the map. I would rather have a game that interacts in a way that doesn't feel like the easiest route, or most common.  If you think about a side scrolling game they really are a bit easier to code, at least the basic functionalities, whatever is added after depends on the creator, however basic movement is just simple, it's almost, or pretty much is taking a dimension out, sure you can jump and have vertical interaction, but that is very short unless making a unique game, also that kind of vertical can be implemented in a top down, you just don't see as large of a gap because the view is different. 

On second thought maybe I will turn the jump into a flying, floating, hovering, levitating type effect. Either that or a sprint/dash or even just a speed or even health boost that last momentartly and is on a timer.

Over all this could probably come in handy as a base line to start a game. I am going to be making classes, and making it object oriented, and pretty much re-formatting the entire thing to go on from here. I just figured it would be easier for me to write it this way first to understand it, as I am no expert on object oriented ideas, and actually quit terrible at them, and don't fully understand all the interactions, at least when I am trying to implement it. I'm sure this will be fine, hopefully make me better at creating classes.

If you would like to work on a game I would gladly make a project and work on this with anyone, even if this is your first time using pygame.  Let me know if you would like to, or if anyone actually uses this, or tries it. I was so happy when the sprites loaded correctly the first time, I had update misspelt and couldn't figure it out, it was pygame.display.upadte and I read over it in both the file and the error Traceback several times without noticing, trying to see where else I used display just to show myself it exists and I wasn't crazy.