This is a third update not any major changes
This version has walking in all 4 directions, as well as animations for those directions (witch will be replaced, trust me it's temporary and all I could find complete. Ignore the tiny sailor moon sprite on the large Zelda/Pokemon/Kirby mashup forest.). I will be doing something with the jump, as this will be top down, and it will not act how it does. Notice when you hold up or down and then click left or right, it is weird and needs to be patched up. See how when you hold left or right you cannot press another direction, that is how each direction should be, I think. I'm gonna play a few good RPGs and see what the movement is like can't remember cause it's something you just do. I may implement another 4 directions, NW, NE, SW, SE. When holding 2 directions, and will have to alter the base code for that. Next upload of this will be big changes to the entire format and way it is programed. I will make sprites when I get time. I usually would have tons of little things to upload, like Tkinter random windows with some functionality, but I have been using my time on this, and researching game development, also some tkinter stuff, but haven't made much of anything important, if I have it was from a guide.