Same Game Now Object Oriented
Had problems with Hard Drive and some other crap with my PC. I didn't work on anything I thought was worth uploading. So I already Missed a day, also had a migrane for over the last 72 hours, however I just updated this little soon to be game, now it is object oriented, or uses a class at least. Needed to make the change as I see this getting a bit larger, who knows maybe I will need to make the change to several files, I hope that is the case, it would be good practice and mean I have a pretty hefty amount of code.

To play you will need python 3.x and to download the .PNG images I have uploaded, they may be in a separate file in this repo, or you can use your own images. Just make sure the background is about 800x800 if you want the screen covered, find a smaller sprite for the player and there is 5 or 6 frames a walk cycle. Match the frame names to the image names in the py file, or just get all the posted images, and play as Sailor Mercury all you want (she is the sprite I think, otherwise known as the Blue one)..

I will be making my own sprites, not sure if I will edit existing sprites or make them from scratch. Same goes for the map, and everything else. I may find a tileset to use, was thinking about Zelda Oracle of Ages/Seasons tile sets. Even remaking a game, just like them, however putting both games on one map. That would be a crazy project. Just making the over world would be good enough, forget the dungeons. If I can and once I have enemies I may do that, then be done, except when I get the urge to work on it, then I can work on a dungeon little by little or something, maybe. Although if I take on something of that magnitude I want to be using Unity, Unreal or something. Even if it's unreal 2004, I have that on an Unreal Tournament game I found, wonder if it's easier than the current version. Without as much things to worry about. Doing this sure makes me want to learn C# and C++, also LUA. In fact i'm sure once I get OOP down in Python, i'll run out of things to do in it and switch.