Another step
In middle of adding a projectile function. A few things I have noted are, when holding left and then pushing right it will not go right, however when holding right and push left it will go left. Same for up and down, when holding down if you push up it will go up, while remaining to hold down, when holding up if you push down (while holding up still) still goes up. This is because it is favoring those directions as they come before the other direction in the loop. Tried to fix this by setting the other directions all to false, did not work.

Another good example that visually demonstates this is if you are on the left or top wall, and hold the other direction, it will rapidly change directions. Anoter issue that needs to be fixed is the invisable wall. On the top and left you cannot go all the way to the wall, this is because rectangles, or positions start at the top left corer. I will have to simply add or subtract half the sprite value from the area it can move. Find the easiest way to do that without changing anything else, I commented where the possible fix will be. After adding the projectile, I may add a sword or attack that is only right infront of the enemy. I will need to then make collision detection and enemies.

I think after reformatting this to use classes, I still think it should be cleaner, or could be. Perhaps using multiple files is the way to go. I would like to research how that works, especially in Python, perhaps even another language like LUA. I imagine a full game should require dictionaries and arrays in separate files, or I guess the term would be, data based programming. That is honestly what makes the most sense to me.