Found on youtube comment and fixed it, random.
Watching a youtube video looking at ways to display user created lists in tkinter, (this video https://www.youtube.com/watch?v=QcXcspHRI3Q). I found some code in the comments or a link, asking why it didn't work, so I fixed it real quick, and that saves me from quickly working on that game, or a Tkinter GUI rushing to get it done to post it here.