Tkinter the possible messy start to something
Possible start to a Dokkan Battle Team organizer. I would like to be able to create teams with being able to see the others. Be sure not to use a unit twice. That way I can map and plan out the virtual Dokkan event teams, and pick 112 units that will all just fit on one given team, making it much easier.