Newest Edit of Game
Will need to find images for the enemy until I upload them witch will be probably the day after this upload, after I get collision and hit boxs done. 
Added an enemy with walk cycle on x axis. Also created the class for enemies that move on y axis, need to change the equation so the y is changed, not x. I will also upload a version with the enemy_verticle class or y moving, once I find sprites for an enemy that is slightly different, or edit the color of the existing, enemy_horizontal/x axis moving.
I will implement hitboxes, probably easy to use auto creating hitboxes, may need some tweaking after.
Still need to add up and down projectiles, unless I make this a firefighting game with houses and stuff on the sides to shoot and put the fire out, that was my creative idea around it lol.